Fixed issue undefined reference to `test_maps_create'
Fixed issue undefined reference to `test_maps_create' - The easiest way
to fix this kind of problem is to right-click on problematic file (in
this case test_maps_create.c and test_maps_create.h) choose option
'remove from this project', and then right-click on name of the project,
choose 'add files', and add this two files again. If this no help that
means there is some bug in the code.