Mountain map generation should be generating correctly now
There is only 1 bug I have found, although it occurs rarely and I can
not reproduce it on purpose. Sometimes the program crashes when running
the mountain generation functions, although most of the times it works.
Each time I run debug every value is within bouds so I don't know what's
causing this. I'll look further into this when I have more time I can
spare.
Also added a function to draw the map on the console! ONLY when drawing,
ONLY WHEN USING THE DRAWING FUNCTION, the 0's shall be considered as
ground! Inside the matrix, 0 is considered as sky/empty :)