Map drawing function completed
Also added a missing ')' to drawing_units.h  [Thanks pawel for spotting
that :) ]
Map drawing works great, and I have added a description on how the map
gets represented on the console.

The program crashing ocasionally when generating mountain map still
persists. I have not yet been able to reproduce this issue, will get to
this later as it's not a critical issue.