save and load the last game
Made by Daniel and Roberto.
Now its possible, when user press "Esc" to save the current game if he
wants to.
Well, when we play the game, if he choose the second option to load the
last game that was exited, the game continues like before, we hope XD
We need to test better to find possible bugs :)