Created missile Terraformer to create a wall when it hits
It is not ready yet to create a wall so for now it behaves like a
regular missile. The goal is to have it finished in a way that you can
only use it when you catch a Terraformer powerup, so sometimes a number
"5" appears, and if you shoot it you get one powerup Terraformer. Then
press 4 to activate it, and wherever the missile falls, it creates a
wall to the top, with the thickness of 3 Xcoordinates.